Created test script and refactored types name
Now the types are called models (to avoid collitions with Python's
standard library). Also created an small README.md with a tutorial in
how to test
