+++ webapp/tests/README.md
@@ -1,0 +1,14 @@
+# Testing webapp
+
+### What's needed in order to test webapp?
+
+- A working installation of RethinkDB
+- Pytest + all the Python dependenceies (venv is highly recommended)
+
+### How can I test it?
+
+Simply run:
+```sh
+cd webapp/tests
+python run.py
+```
